Refatorado, corrigido alguns erros.

+++ src/main/java/br/com/alelo/consumer/consumerpat/domain/service/CardContext.java
@@ -3,6 +3,8 @@ package br.com.alelo.consumer.consumerpat.domain.service;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 
 @Builder
 @NoArgsConstructor
@@ -13,11 +15,11 @@ public class CardContext {
         this.strategy = strategy;
     }
 
-    public Double creditValue(Double balance, Double value){
+    public BigDecimal creditValue(BigDecimal balance, BigDecimal value){
     return strategy.creditValue(balance, value);
     }
 
-    public Double applyCashback(Double balance, Double value){
+    public BigDecimal applyCashback(BigDecimal balance, BigDecimal value){
         return strategy.applyCashback(balance, value);
     }
  }
