+++ src/main/java/br/com/alelo/consumer/consumerpat/integration/rest/controller/dto/request/CardBalanceRequestPostV1.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CardBalanceRequestPostV1 {
     private Long number;
-    private Double value;
+    private BigDecimal value;
 }
