+++ src/test/java/br/com/alelo/consumer/consumerpat/helper/ExtractHelper.java
@@ -3,12 +3,13 @@ package br.com.alelo.consumer.consumerpat.helper;
 import br.com.alelo.consumer.consumerpat.domain.entity.Card;
 import br.com.alelo.consumer.consumerpat.domain.entity.Extract;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
 public class ExtractHelper {
 
-    public static Extract buildExtract(Card card, Double value) {
+    public static Extract buildExtract(Card card, BigDecimal value) {
         return Extract.builder()
                 .cards(Set.of(card))
                 .productDescription("alimentos")
